Create rockpaperscissors.py
The object of the rock-paper-scissor python project is to build a game for a single player that plays with a computer, anywhere, and anytime. This project is based on the rules:

1. Rock blunts scissors so rock wins
2. Scissors cut the paper so scissors win
3. Paper cover rock so paper wins

This project is built using random modules, and the basic concept of python.
In this python project, players have to choose anyone from rock, paper, and scissors. Then clicking on the play button will show the result of the game.